requirements.txt
Flask==3.1.*
python-frontmatter==1.1.*
markdown==3.7.*
Pygments==2.19.*
